Final directory restructure
