+++ arches_project/arches_project.py
@@ -196,7 +196,10 @@ class ArchesProject:
     def initGui(self):
         """Create the menu entries and toolbar icons inside the QGIS GUI."""
 
-        icon_path = ':/plugins/arches_project/icons/arches.png'
+        icon_path = os.path.join(
+            self.plugin_dir,
+            'icons',
+            'arches.png')
         self.add_action(
             icon_path,
             text=self.tr(u'Arches Project'),
