+++ arches_project/pb_tool.cfg
@@ -60,7 +60,7 @@ compiled_ui_files:
 resource_files: resources.qrc
 
 # Other files required for the plugin
-extras: arches_project/metadata.txt icons/arches.png
+extras: ../metadata.txt icons/arches.png
 
 # Other directories to be deployed with the plugin.
 # These must be subdirectories under the plugin directory
