+++ arches_project/resources.py
@@ -234,7 +234,7 @@ qt_resource_struct_v2 = b"\
 \x00\x00\x00\x36\x00\x02\x00\x00\x00\x01\x00\x00\x00\x04\
 \x00\x00\x00\x00\x00\x00\x00\x00\
 \x00\x00\x00\x46\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01\x8b\xf1\xa6\x4c\x5f\
+\x00\x00\x01\x8f\x7e\x47\x2d\x22\
 "
 
 qt_version = [int(v) for v in QtCore.qVersion().split('.')]
