+++ arches_project/ui/arches_project_dialog_base.ui
@@ -18,7 +18,7 @@
   </property>
   <property name="windowIcon">
    <iconset>
-    <normaloff>icons/arches_transparent_bg.png</normaloff>icons/arches_transparent_bg.png</iconset>
+    <normaloff>../icons/arches_transparent_bg.png</normaloff>../icons/arches_transparent_bg.png</iconset>
   </property>
   <property name="modal">
    <bool>false</bool>
